requirements.txt
streamlit
torch
torchvision
transformers
pillow
pandas
pyyaml
numpy